Bug 1581058 - Move RenderTaskCache into its own module. r=gw
Differential Revision: https://phabricator.services.mozilla.com/D45855

[wrupdater] From https://hg.mozilla.org/mozilla-central/rev/362feebe585e002ecbfa1ac8938092be02320856
