+++ webrender/src/glyph_cache.rs
@@ -5,7 +5,7 @@
 use crate::glyph_rasterizer::{FontInstance, GlyphFormat, GlyphKey, GlyphRasterizer};
 use crate::internal_types::FastHashMap;
 use crate::render_backend::{FrameId, FrameStamp};
-use crate::render_task::RenderTaskCache;
+use crate::render_task_cache::RenderTaskCache;
 use crate::resource_cache::ResourceClassCache;
 use std::sync::Arc;
 use crate::texture_cache::{EvictionNotice, TextureCache};
